charms/candid/requirements.txt
ops >= 1.4.0
ops-lib-pgsql